@@ -9,48 +9,54 @@ package wba.citta.cognitivedistance;
 import java.util.*;
 
 /**
- *	状態の履歴を保持するクラスです。
- *	コンストラクタの引数で設定された数だけ、ノードIDを保持します。
- *	最大の要素数を越えた場合、古い順に削除されます。
+ *    状態の履歴を保持するクラスです。
+ *    コンストラクタの引数で設定された数だけ、ノードIDを保持します。
+ *    最大の要素数を越えた場合、古い順に削除されます。
  */
-public class StateBuffer extends LinkedList {
+public class StateBuffer extends ArrayList<Integer> {
 
-	/* 保持する最大の要素数 */
-	private int maxCDLngth;
+    /**
+     * 
+     */
+    private static final long serialVersionUID = 1L;
+    /* 保持する最大の要素数 */
+    private int maxCDLngth;
 
-	/**
-	 * コンストラクタ
-	 * @param int maxCDLngth 保持する最大の要素数
-	 */
-	public StateBuffer(int maxCDLngth) {
-		this.maxCDLngth = maxCDLngth;
-	}
+    /**
+     * コンストラクタ
+     * @param int maxCDLngth 保持する最大の要素数
+     */
+    public StateBuffer(int maxCDLngth) {
+        this.maxCDLngth = maxCDLngth;
+    }
 
 
-	/**
-	 *	ノードIDをリストに追加します。
-	 *	@param Integer stateID ノードID
-	 */
-	public void add(Integer stateID){
-		/* リストに追加 */
-		super.add(stateID);
-		/* リストのサイズをMAXSIZEに調整 */
-		if(size() > maxCDLngth){
-			removeFirst();
-		}
-	}
+    /**
+     *    ノードIDをリストに追加します。
+     *    @param Integer stateID ノードID
+     * @return 
+     */
+    public boolean add(Integer stateID){
+        /* リストに追加 */
+        super.add(stateID);
+        /* リストのサイズをMAXSIZEに調整 */
+        if(size() > maxCDLngth){
+            this.remove(0);
+        }
+        return true;
+    }
 
-	/* リスト内の要素を表示 */
+    /* リスト内の要素を表示 */
 // 2001.05.25 削除 miyamoto 古いバージョンのjavaに対応
 // LinkedListをVectorに変更したためtoStringをオーバーライドできない(必要ない)
-//	public String toString() {
-//		StringBuffer sb = new StringBuffer();
-//		ListIterator li = listIterator();
-//		while(li.hasNext()) {
-//			Integer id = (Integer)li.next();
-//			sb.append(id + ",");
-//		}
-//		return sb.toString();
-//	}
+//    public String toString() {
+//        StringBuffer sb = new StringBuffer();
+//        ListIterator li = listIterator();
+//        while(li.hasNext()) {
+//            Integer id = (Integer)li.next();
+//            sb.append(id + ",");
+//        }
+//        return sb.toString();
+//    }
 }
 
