@@ -6,251 +6,330 @@
  */
 package wba.citta.doorkeydemo;
 
+import java.awt.Color;
 import java.io.*;
 import java.util.*;
-import wba.citta.gsa.*;
 
+import javax.swing.JFrame;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import ch.qos.logback.classic.LoggerContext;
+import ch.qos.logback.classic.filter.ThresholdFilter;
+import ch.qos.logback.classic.spi.ILoggingEvent;
+import ch.qos.logback.core.Appender;
+
+import wba.citta.CittaRunner;
+import wba.citta.IterationEvent;
+import wba.citta.IterationEventListener;
+import wba.citta.StepCountViewer;
+import wba.citta.StepEvent;
+import wba.citta.LogViewer;
+import wba.citta.StepEventListener;
+import wba.citta.doorkeydemo.environment.Disposition;
+import wba.citta.doorkeydemo.environment.Environment;
+import wba.citta.gsa.*;
+import wba.citta.gsa.viewer.AgentViewer;
+import wba.citta.gsa.viewer.SharedMemoryViewer;
+import wba.citta.gsa.viewer.TreeViewer;
+import wba.citta.gui.ViewerPanel;
+import wba.citta.gui.ViewerPanelContainer;
+import wba.citta.gui.ViewerPanelEvent;
+import wba.citta.util.EventPublisherSupport;
 
 /**
  * Citta
  */
-public class DoorKeyDemo {
+public class DoorKeyDemo extends JFrame implements CittaRunner {
+    @SuppressWarnings("serial")
+    public static class ApplicationException extends RuntimeException {
+        public ApplicationException() {
+            super();
+        }
 
-	/*  */
-	private EnvironmentAgent environmentAgent = null;
-	/* Citta */
-	private GSA gsa = null;
+        public ApplicationException(String message, Throwable cause,
+                boolean enableSuppression, boolean writableStackTrace) {
+            super(message, cause, enableSuppression, writableStackTrace);
+        }
 
-	/*  */
-	private DemoProperty prop = null;
+        public ApplicationException(String message, Throwable cause) {
+            super(message, cause);
+        }
 
-	/*  */
-	private int saveCount;
-	private String saveFileName;
-	private String loadFileName;
-	private final int TIME_OUT_STEP;
-	private final int SLEEP_TIME;
+        public ApplicationException(String message) {
+            super(message);
+        }
 
-	/*  */
-	private Vector goal = null;
+        public ApplicationException(Throwable cause) {
+            super(cause);
+        }        
+    }
 
+    private static final Logger logger = LoggerFactory.getLogger(DoorKeyDemo.class);
+    private static final long serialVersionUID = 1L;
+    /*  */
+    private EnvironmentAdapter environmentAdapter = null;
+    /* Citta */
+    private GSA gsa = null;
 
-	////////////////////////////////////////////////////////////////
-	// 
+    /*  */
+    private DemoProperty props = null;
 
-	/**
-	 * 
-	 */
-	public DoorKeyDemo(String propFileName) throws Exception {
+    /*  */
+    private int saveCount;
+    private String saveFileName;
+    private String loadFileName;
+    private final int maxStepCount;
+    private final int sleepTime;
 
-		/*  */
-		try {
-			prop = new DemoProperty(propFileName);
-		}catch(Exception e) {
-			e.printStackTrace();
-			System.exit(0);
-		}
+    private int stepCount = 0;
+    private int toGoalStepCount = 0;
+    private int iteration = 0;
+    private int goalCount = 0;
 
-		gsa = new GSA(prop.getGSAPropFileName());
+    private List<Integer> goal = null;
+    private List<Integer> state = null;
+    private List<Integer> subgoal = null;
 
-		environmentAgent = new EnvironmentAgent(prop.getEnvFileName(), 
-		         prop.getDoorOpenMode(), 1/**/);
+    private ViewerPanelContainer container;
+    
+    private final EventPublisherSupport<StepEvent, StepEventListener> stepEventListeners = new EventPublisherSupport<>(StepEvent.class, StepEventListener.class);
+    private final EventPublisherSupport<IterationEvent, IterationEventListener> iterationEventListeners = new EventPublisherSupport<>(IterationEvent.class, IterationEventListener.class);
 
-		/*  */
-		saveCount = prop.getSaveStepNum();
-		saveFileName = prop.getSaveFileName();
-		loadFileName = prop.getLoadFileName();
-		TIME_OUT_STEP = prop.getTimeOutStepNum();
-		SLEEP_TIME = prop.getSleepTime();
+    public DoorKeyDemo(String propFileName) throws Exception {
+        props = new DemoProperty(propFileName);
+        saveCount = props.getSaveStepNum();
+        saveFileName = props.getSaveFileName();
+        loadFileName = props.getLoadFileName();
+        maxStepCount = props.getTimeOutStepNum();
+        sleepTime = props.getSleepTime();
+        initComponents();
+        initEnvironmentAdapter();
+        initLogViewer();
+        initStepCountViewer();
+        initGoal();
+        initGSA();
+        firePanelPopulated();
+    }
 
-		/*  */
-		if(!loadFileName.equals("")) {
-			gsa.load(loadFileName);
-		}
+    private void initComponents() {
+        setDefaultCloseOperation(EXIT_ON_CLOSE);
+        setTitle(getClass().getName());
+        container = new ViewerPanelContainer(this);
+        getContentPane().add(container);
+        setSize(720, 500);
+        setVisible(true);
+    }
 
-		/* gsa */
-		initGoal();
-		setGoal();
+    private void initGSA() throws IOException {
+        final GSAProperty gsaProps = new GSAProperty(props.getGSAPropFileName());
+        final GSAFactory gsaFactory = new GSAFactory();
+        gsaFactory.addGSAFactoryEventListener(new GSAFactoryEventListener() {
+            @Override
+            public void gsaCreated(GSAFactoryEvent evt) {
+                final GSA gsa = evt.getGSA();
+                if (gsaProps.isShowAgentViewer()) {
+                    AgentViewer agentViewer = new AgentViewer(gsaProps.getColorTable(), Color.GRAY);
+                    agentViewer.bind(gsa);
+                    agentViewer.setSize(gsaProps.getAgentViewerInitSize());
+                    firePanelCreated(agentViewer);
+                }
+                if (gsaProps.isShowGoalStackViewer()) {
+                    SharedMemoryViewer sharedMemoryViewer = new SharedMemoryViewer(gsaProps.getColorTable());
+                    sharedMemoryViewer.bind(gsa.getSharedMemory());
+                    sharedMemoryViewer.setSize(gsaProps.getSharedMemoryViewerInitSize());
+                    firePanelCreated(sharedMemoryViewer);
+                }
+                if (gsaProps.isShowFailAgentTreeViewer()) {
+                    TreeViewer treeViewer = new TreeViewer(gsaProps.getColorTable());
+                    treeViewer.bind(gsa.getFailAgentTree());
+                    firePanelCreated(treeViewer);
+                }
+            }
+        });
+        gsaFactory.populateWithGSAProperty(gsaProps);
+        gsa = gsaFactory.createGSA();
+        
+        if(!loadFileName.equals("")) {
+            gsa.load(loadFileName);
+        }
 
-		// 
-		initLogFile("Step.log");
-
-	}
-
-	/**
-	 * 
-	 */
-	private void initGoal() {
-		goal = new Vector();
-		goal.add(null);
-		goal.add(null);
-		goal.add(null);
-		goal.add(new Integer(1));
-		goal.add(null);
-		goal.add(null);
-		goal.add(null);
-	}
-
-	////////////////////////////////////////////////////////////////
-	// public
-
-	/**
-	 * 
-	 */
-	public void repeatProcess() {
-
-		int toGoalStepCount = 0; /*  */
-		int stepCount = 0;       /*  */
-
-		/*
-		 * 
-		 * xyID
-		 */
-		Vector state = null;    /*  */
-		Vector subgoal = null;  /* CITTA */
-
-		/* GSA */
-		while(true) {
-
-			/*   */
-			try {
-				Thread.sleep(SLEEP_TIME);
-			}catch(InterruptedException e) {
-				System.out.println(e);
-			}
-
-
-			/*    */
-			if( isReachGoal(state, goal) || toGoalStepCount==TIME_OUT_STEP ) {
-
-				/*  */
-				System.out.println("++++++++++++++++++++++++++++++++++++++++");
-				System.out.println("   " + toGoalStepCount);
-				logOut(toGoalStepCount);
-				System.out.println("++++++++++++++++++++++++++++++++++++++++");
-
-				toGoalStepCount = 0;
-				restart();
-				subgoal = null;
-
-				if(isReachGoal(state, goal) ) {
-					goalReachProcess();
-				}
-			}
-
-			/*  */
-			toGoalStepCount++;
-			stepCount++;
-			System.out.println("");
-			System.out.println(" step count " + stepCount);
-
-			/*  */
-			/*
-			 * CITTA 
-			 * 
-			 */
-			int action = -1;
-			if(subgoal != null) {
-				/* 3action */
-				if(subgoal.get(2) != null) {
-					action = ((Integer)subgoal.get(2)).intValue();
-				}
-			}
-			state = environmentAgent.move(action);
-
-			/* CITTA */
-			/*  */
-			if(stepCount == saveCount) {
-				if(!saveFileName.equals("")) {
-					gsa.save(saveFileName);
-				}
-			}
-			/* state */
-			subgoal = gsa.exec(state);
-
-			/*  */
-//			gsa.printStack();
-//			gsa.printGoalTree();
-
-		}
-	}
+        gsa.setGoal(goal);
+        fireIterationStartedEvent();
+    }
 
 
-	////////////////////////////////////////////////////////////////
-	// private
+    private void initEnvironmentAdapter() throws IOException {
+        final String fileName = props.getEnvFileName();
+        logger.info(String.format("Loading map file %s", fileName));
+        final Environment environment = new Environment(Disposition.load(fileName));
+        environmentAdapter = new EnvironmentAdapter(
+            environment, 
+            props.getDoorOpenMode(),
+            1
+        );
+        addIterationEventListener(environment);
+        firePanelCreated(environment);
+    }
 
-	/**
-	 * 
-	 * (null)Statetrue
-	 */
-	private boolean isReachGoal(Vector state, Vector goal) {
+    private void initLogViewer() {
+        final ViewerPanel logViewer = new LogViewer();
+        firePanelCreated(logViewer);
+    }
 
-		if(goal == null || state == null) {
-			return false;
-		}
+    private void initStepCountViewer() {
+        final StepCountViewer stepCountViewer = new StepCountViewer();
+        addStepEventListener(stepCountViewer);
+        addIterationEventListener(stepCountViewer);
+        firePanelCreated(stepCountViewer);
+    }
 
-		for(int i = 0; i < state.size(); i++) {
-			Integer sElement = (Integer)state.get(i);
-			Integer gElement = (Integer)goal.get(i);
-			if(gElement != null && !sElement.equals(gElement)) {
-				return false;
-			}
-		}
-		return true;
-	}
+    private void initGoal() {
+        goal = Arrays.asList(null, null, null, 1, null, null, null);
+    }
 
-	/**
-	 * 
-	 */
-	private void goalReachProcess() {
-		/*  */
-		environmentAgent.goalReach();
-	}
+    protected void firePanelCreated(ViewerPanel panel) {
+        container.panelCreated(new ViewerPanelEvent(this, panel));
+    }
 
-	/**
-	 * 
-	 */
-	private void restart() {
-		gsa.reset();
-		environmentAgent.restart();
-		/* gsa */
-		setGoal();
-	}
+    protected void firePanelPopulated() {
+        container.panelPopulated(new ViewerPanelEvent(this));
+    }
 
-	/**
-	 * CITTA
-	 */
-	private void setGoal() {
-		gsa.setGoal(goal);
-	}
+    public void step() {
+        if (isGoalReached(state, goal)) {
+            logger.info(String.format("Goal reached (total step count: %d)", toGoalStepCount));
+            goalCount++;
+            fireIterationEndedEvent(true);
+            restart();
+        } else if (toGoalStepCount >= maxStepCount) {
+            logger.info(String.format("Step count reached the maximum step count (step count: %d)", toGoalStepCount));
+            fireIterationEndedEvent(false);
+            restart();
+        }
 
+        toGoalStepCount++;
+        stepCount++;
 
-	//////////////////////////////////////////////////////////////
-	// 
+        if (logger.isTraceEnabled())
+            logger.trace(String.format("iteration: %d", stepCount));
+        fireStepEvent();
 
-	private PrintWriter printWriter; // 
+        int action = -1;
+        if (subgoal != null) {
+            /* 3action */
+            if(subgoal.get(2) != null) {
+                action = ((Integer)subgoal.get(2)).intValue();
+            }
+        }
+        state = environmentAdapter.move(action);
 
-	private void initLogFile(String fileName) {
-		try{
-			FileOutputStream fileOutputStream = new FileOutputStream(fileName,
-			        false);
-			printWriter = new PrintWriter(fileOutputStream, true);
-		}catch(Exception e) {
-		}
-	}
+        if (stepCount == saveCount) {
+            if (!saveFileName.equals("")) {
+                try {
+                    gsa.save(saveFileName);
+                } catch (IOException e) {
+                    throw new ApplicationException(e);
+                }
+            }
+        }
+        /* state */
+        subgoal = gsa.exec(state);
 
-	private void logOut(int stepNum) {
-		printWriter.println(stepNum);
-	}
+    }
 
+    public void repeatProcess() {
+        while (true) {
+            try {
+                Thread.sleep(sleepTime);
+            } catch(InterruptedException e) {
+                throw new ApplicationException(e);
+            }
+            step();
+        }
+    }
 
-	////////////////////////////////////////////
-	//  
+    private void fireIterationStartedEvent() {        
+        iterationEventListeners.fire("iterationStarted", new IterationEvent(this, iteration, goalCount, false));
+    }
 
-	public static void main(String[] args) throws Exception {
-		DoorKeyDemo dkDemo = new DoorKeyDemo(args[0]);
-		dkDemo.repeatProcess();
-	}
+    private void fireIterationEndedEvent(boolean achieved) {        
+        iterationEventListeners.fire("iterationEnded", new IterationEvent(this, iteration, goalCount, achieved));
+    }
 
+    private void fireStepEvent() {
+        stepEventListeners.fire("nextStep", new StepEvent(this, stepCount, toGoalStepCount));
+    }
 
+    private boolean isGoalReached(List<Integer> state, List<Integer> goal) {
+        if(goal == null || state == null) {
+            return false;
+        }
+
+        for(int i = 0; i < state.size(); i++) {
+            Integer sElement = (Integer)state.get(i);
+            Integer gElement = (Integer)goal.get(i);
+            if(gElement != null && !sElement.equals(gElement)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    /*
+    private void reset() {
+        stepCount = 0;
+        iteration = 0;
+        goalCount = 0;
+        restart();
+    }
+    */
+
+    private void restart() {
+        gsa.reset();
+        gsa.setGoal(goal);
+        environmentAdapter.restart();
+        toGoalStepCount = 0;
+        state = null;
+        subgoal = null;
+        iteration++;
+        fireIterationStartedEvent();
+    }
+
+    private static void setupLogging() {
+        final LoggerContext ctx = (LoggerContext)LoggerFactory.getILoggerFactory();
+        final ch.qos.logback.classic.Logger rootLogger = ctx.getLogger(Logger.ROOT_LOGGER_NAME);
+        final Appender<ILoggingEvent> appender = LogViewer.getAppender();
+        appender.setContext(ctx);
+        final ThresholdFilter filter = new ThresholdFilter();
+        filter.setContext(ctx);
+        filter.setLevel("INFO");
+        filter.start();
+        appender.addFilter(filter);
+        rootLogger.addAppender(appender);
+        appender.start();
+    }
+
+    public void addStepEventListener(StepEventListener listener) {
+        stepEventListeners.addEventListener(listener);
+    }
+
+    public void removeStepEventListener(StepEventListener listener) {
+        stepEventListeners.removeEventListener(listener);
+    }
+
+    public void addIterationEventListener(IterationEventListener listener) {
+        iterationEventListeners.addEventListener(listener);
+    }
+
+    public void removeIterationEventListener(IterationEventListener listener) {
+        iterationEventListeners.removeEventListener(listener);
+    }
+
+    public static void main(String[] args) throws Exception {
+        setupLogging();
+        DoorKeyDemo dkDemo = new DoorKeyDemo(args[0]);
+        dkDemo.repeatProcess();
+    }
 }
