@@ -14,93 +14,87 @@ import java.util.*;
  */
 public class LogAgent extends Agent {
 
-	private final String LOG_FILE_NAME = "AgentLog_";
+    private final String LOG_FILE_NAME = "AgentLog_";
 
-	private FileInputStream fileInputStream = null;
-	private ObjectInputStream objectInputStream = null;
+    private FileInputStream fileInputStream = null;
+    private ObjectInputStream objectInputStream = null;
 
-	/**
-	 * コンストラクタ
-	 * @param int agid  エージェントID
-	 * @param SharedMemory sharedMemory  state・goalを管理する共有メモリ
-	 * @param boolean[] useNode  ノードの使用、不使用を設定した配列
-	 */
-	public LogAgent(int agid, boolean[] useNode, SharedMemory sharedMemory) {
-		super(agid, useNode, sharedMemory);
+    /**
+     * コンストラクタ
+     * @param int agid  エージェントID
+     * @param SharedMemory sharedMemory  state・goalを管理する共有メモリ
+     * @param boolean[] useNode  ノードの使用、不使用を設定した配列
+     */
+    public LogAgent(int agid, boolean[] useNode, SharedMemory sharedMemory) {
+        super(agid, useNode, sharedMemory);
 
-		/* ログの読み込みストリーム生成 */
-		String fileName = LOG_FILE_NAME + agid + ".log";
-		try{
-			fileInputStream = new FileInputStream(fileName);
-			objectInputStream = new ObjectInputStream(fileInputStream);
-		}catch(Exception e){
-			System.out.println(e);
-			System.exit(0);
-		}
+        /* ログの読み込みストリーム生成 */
+        String fileName = LOG_FILE_NAME + agid + ".log";
+        try{
+            fileInputStream = new FileInputStream(fileName);
+            objectInputStream = new ObjectInputStream(fileInputStream);
+        }catch(Exception e){
+            System.out.println(e);
+            System.exit(0);
+        }
 
-	}
+    }
 
 
-	///////////////////////////////////////////////////////////////////////
-	// public
+    ///////////////////////////////////////////////////////////////////////
+    // public
 
-	/**
-	 * Agentクラスを継承して作成しているため、形式的に実装<BR>
-	 */
-	public void learn(Vector state, boolean flagGoalReach, double profit) {
-	}
+    /**
+     * Agentクラスを継承して作成しているため、形式的に実装<BR>
+     */
+    public void learn(List<Integer> state, boolean flagGoalReach, double profit) {
+    }
 
-	/**
-	 * エージェント固有の実行処理を行ないます。<BR>
-	 * ログファイルから情報を取得し、取得した情報をサブゴールとして出力します。
-	 * <BR>
-	 * @param Vector state 現在の状態
-	 * @param Vector goalElementArray GoalStackElementのVector
-	 * @return Vector サブゴール
-	 */
-	public Vector execProcess(Vector state, Vector goalElement) {
+    /**
+     * エージェント固有の実行処理を行ないます。<BR>
+     * ログファイルから情報を取得し、取得した情報をサブゴールとして出力します。
+     * <BR>
+     * @param Vector state 現在の状態
+     * @param Vector goalElementArray SharedMemory.GoalStackElementのVector
+     * @return Vector サブゴール
+     */
+    @SuppressWarnings("unchecked")
+    public List<Integer> execProcess(List<Integer> state, List<SharedMemory.GoalStackElement> goalElement) {
+        /* ログの読み込み 出力 */
+        List<Integer> nextState = null;
+        try{
+            nextState = (List<Integer>)objectInputStream.readObject();
+        } catch (IOException e) {
+            throw new GSAException(e);
+        } catch (ClassNotFoundException e){
+            throw new GSAException(e);
+        }
+        return nextState;
 
-//		System.out.println("   state:" + state);
-//		System.out.println("   goal :" + goalValue);
+    }
 
-		/* ログの読み込み 出力 */
-		Vector nextState = null;
-		try{
-			nextState = (Vector)objectInputStream.readObject();
-		}catch(Exception e){
-			System.out.println("AGID " + AGID);
-			System.out.println(e);
-		}
+    /**
+     * Agentクラスを継承して作成しているため、形式的に実装<BR>
+     */
+    public void reset() {
+    }
 
-//		System.out.println("   next state :" + nextState);
+    /**
+     * Agentクラスを継承して作成しているため、形式的に実装<BR>
+     */
+    public void suspend() {
+    }
 
-		return nextState;
+    /**
+     * Agentクラスを継承して作成しているため、形式的に実装<BR>
+     */
+    public void save(String fileName) { 
+    }
 
-	}
-
-	/**
-	 * Agentクラスを継承して作成しているため、形式的に実装<BR>
-	 */
-	public void reset() {
-	}
-
-	/**
-	 * Agentクラスを継承して作成しているため、形式的に実装<BR>
-	 */
-	public void suspend() {
-	}
-
-	/**
-	 * Agentクラスを継承して作成しているため、形式的に実装<BR>
-	 */
-	public void save(String fileName) { 
-	}
-
-	/**
-	 * Agentクラスを継承して作成しているため、形式的に実装<BR>
-	 */
-	public void load(String fileName) {
-	}
-
+    /**
+     * Agentクラスを継承して作成しているため、形式的に実装<BR>
+     */
+    public void load(String fileName) {
+    }
 }
 
