@@ -16,69 +16,67 @@ import java.util.*;
  */
 public abstract class LogOutAgent extends Agent {
 
-	private final String LOG_FILE_NAME = "AgentLog_";
+    private final String LOG_FILE_NAME = "AgentLog_";
 
-	private FileOutputStream fileOutputStream = null;
-	private ObjectOutputStream objectOutputStream = null;
-
-
-	/**
-	 * コンストラクタ
-	 * @param int agid  エージェントID
-	 * @param boolean[] useNode  ノードの使用、不使用を設定した配列
-	 * @param SharedMemory sharedMemory  state・goalを管理する共有メモリ
-	 */
-	public LogOutAgent(int agid, boolean[] useNode,
-	         SharedMemory sharedMemory) {
-		super(agid, useNode, sharedMemory);
-		initLogFile(agid + ".log");
-	}
-
-	/**
-	 * ログファイルの初期化
-	 * 引数で指定されたファイル名で、ログファイルを生成します。
-	 * @param String fileName ファイル名
-	 */
-	private void initLogFile(String fileName) {
-
-		String logFileName = LOG_FILE_NAME + fileName;
-
-		try {
-			/* ストリームの作成 */
-			fileOutputStream = new FileOutputStream(logFileName, false);
-			objectOutputStream = new ObjectOutputStream(fileOutputStream);
-		}catch(Exception e) {
-			e.printStackTrace();
-			System.out.println(e);
-			System.exit(0);
-		}
-
-	}
+    private FileOutputStream fileOutputStream = null;
+    private ObjectOutputStream objectOutputStream = null;
 
 
-	///////////////////////////////////////////////////////////////////////
-	// public
+    /**
+     * コンストラクタ
+     * @param int agid  エージェントID
+     * @param boolean[] useNode  ノードの使用、不使用を設定した配列
+     * @param SharedMemory sharedMemory  state・goalを管理する共有メモリ
+     */
+    public LogOutAgent(int agid, boolean[] useNode,
+             SharedMemory sharedMemory) {
+        super(agid, useNode, sharedMemory);
+        initLogFile(agid + ".log");
+    }
 
-	/**
-	 * Agentクラスのexecをオーバーライドし、サブゴールをログとしてファイル出力
-	 * する処理を追加。<BR>
-	 * @param Vector state
-	 * @param Vector goalElementArray
-	 * @return subgoal
-	 */
-	public Vector exec(Vector state, Vector goalElementArray) {
+    /**
+     * ログファイルの初期化
+     * 引数で指定されたファイル名で、ログファイルを生成します。
+     * @param String fileName ファイル名
+     */
+    private void initLogFile(String fileName) {
 
-		Vector subgoal = execProcess(state, goalElementArray);
+        String logFileName = LOG_FILE_NAME + fileName;
 
-		try {
-			objectOutputStream.writeObject(subgoal);
-		}catch(Exception e) {
-			System.out.println(e);
-		}
+        try {
+            /* ストリームの作成 */
+            fileOutputStream = new FileOutputStream(logFileName, false);
+            objectOutputStream = new ObjectOutputStream(fileOutputStream);
+        }catch(Exception e) {
+            e.printStackTrace();
+            System.out.println(e);
+            System.exit(0);
+        }
 
-		return subgoal; 
-	}
+    }
 
 
+    ///////////////////////////////////////////////////////////////////////
+    // public
+
+    /**
+     * Agentクラスのexecをオーバーライドし、サブゴールをログとしてファイル出力
+     * する処理を追加。<BR>
+     * @param Vector state
+     * @param Vector goalElementArray
+     * @return subgoal
+     */
+    public List<Integer> exec(List<Integer> state, List<SharedMemory.GoalStackElement> goalElementArray) {
+
+        List<Integer> subgoal = execProcess(state, goalElementArray);
+
+        try {
+            objectOutputStream.writeObject(subgoal);
+        }catch(Exception e) {
+            System.out.println(e);
+        }
+
+        return subgoal; 
+    }
 }
 
