@@ -7,7 +7,8 @@
 package wba.citta.gsa;
 
 import java.util.*;
-import wba.citta.gsa.viewer.*;
+
+import wba.citta.util.EventPublisherSupport;
 
 /**
  * State・Goalを管理する共有メモリです。<BR>
@@ -15,190 +16,199 @@ import wba.citta.gsa.viewer.*;
  * 共有メモリへの操作は、基本的にメソッドの引数で指定されたindexの要素ごとに
  * 行ないます。
  */
-public class SharedMemory {
+public class SharedMemory {   
+    public class GoalStack {
+        public LinkedList<GoalStackElement>[] goalStackArray;
 
-	/* State用 Integerの配列 */
-	private Integer[] stateArray = null;
+        @SuppressWarnings("unchecked")
+        public GoalStack(int size) {
+            goalStackArray = new LinkedList[size];
+            for(int i = 0; i < size; i++) {
+                goalStackArray[i] = new LinkedList<GoalStackElement>();
+            }
+        }
 
-	/* Goal用 LinkedListの配列 LinkedListにはElementを設定 */
-	private LinkedList[] goalStackArray = null;
+        /**
+         * 現在のゴールスタックを取得します
+         */
+        public List<GoalStackElement> getGoalStackForNode(int i) {
+            return Collections.unmodifiableList(goalStackArray[i]);
+        }
 
-	/**
-	 * 共有メモリのノード数
-	 */
-	public final int LENGTH;
+        /**
+         * Goalの指定された位置(ノード)の値をスタックからGETで取得します。
+         * @param int index
+         * @return GoalStackElement  ゴールの要素
+         */
+        public GoalStackElement getGoalForNode(int index) {
+            GoalStackElement elm = null;
+            if (goalStackArray[index].size() > 0) {
+                elm = (GoalStackElement)goalStackArray[index].getLast();
+            }
+            return elm;
+        }
 
-	/* ゴールスタックの状態を表示するviewer */
-	private SharedMemoryViewer viewer;
+        /**
+         * 指定されたゴールの要素をGoalの指定された位置(ノード)のスタックにPUSHで
+         * 設定します。       
+         * @param int index
+         * @param GoalStackElement elm ゴールの要素
+         */
+        public void pushGoal(int index, GoalStackElement elm) {
+            goalStackArray[index].add(elm);
+            fireSharedMemoryChanged();
+        }
 
-	/**
-	 * コンストラクタ
-	 * @param int size ノード数（配列のサイズ）
-	 * @param boolean isShowViewer スタックの状態をグラフィック表示するかどうか
-	 */
-	public SharedMemory(int size, boolean isShowViewer) {
-		LENGTH = size;
-		stateArray = new Integer[LENGTH];
-		goalStackArray = new LinkedList[LENGTH];
-		for(int i = 0; i < LENGTH; i++) {
-			goalStackArray[i] = new LinkedList();
-		}
+        /**
+         * Goalの指定された位置(ノード)の値をスタックから削除します。
+         * @param index 
+         */
+        public void removeGoal(int index, int agid) {
+            if (goalStack.goalStackArray[index].getLast().agid != agid)
+                throw new GSAException("WTF?");
+            goalStackArray[index].removeLast();
+            fireSharedMemoryChanged();
+        }
 
-		if(isShowViewer == true) {
-			viewer = new SharedMemoryViewer(stateArray, goalStackArray);
-		}
-	}
+        /**
+         * Goalの要素を全てクリアします。
+         */
+        public void removeAllGoal() {
+            for(int i = 0; i < size; i++) {
+                goalStack.goalStackArray[i].clear();
+            }
+            fireSharedMemoryChanged();
+        }
 
+        /**
+         * 全ノードのゴールをVectorで取得します。
+         * @param Vector GoalValueのVector
+         */
+        public List<Integer> getGoalValueArray() {
+            List<Integer> goal = new ArrayList<Integer>();
+            for(int i = 0; i < size; i++) {
+                GoalStackElement goalElement = goalStack.getGoalForNode(i);
+                if(goalElement != null) {
+                    goal.add(goalElement.value);
+                }else {
+                    goal.add(null);
+                }
+            }
+            return goal;
+        }
+    }
 
-	/**
-	 * Stateの指定された位置(ノード)の値を取得します。
-	 * @param int index 
-	 * @return Integer  
-	 */
-	public Integer getState(int index) {
-		return stateArray[index];
-	}
+    public class Latch {
+        public Integer[] stateArray;
 
-	/**
-	 * 指定された値をStateの指定された位置(ノード)に設定します。
-	 * @param int index 
-	 * @param Integer value 
-	 */
-	public void setState(int index, Integer value) {
-		stateArray[index] = value;
-	}
+        public Latch(int size) {
+            this.stateArray = new Integer[size];
+        }
 
-	/**
-	 * Goalの指定された位置(ノード)の値をスタックからGETで取得します。
-	 * @param int index
-	 * @return GoalStackElement  ゴールの要素
-	 */
-	public GoalStackElement getGoal(int index) {
-		GoalStackElement elm = null;
-		if(goalStackArray[index].size() > 0) {
-			elm = (GoalStackElement)goalStackArray[index].getLast();
-		}
-		return elm;
-	}
+        /**
+         * Stateの指定された位置(ノード)の値を取得します。
+         * @param int index 
+         * @return Integer  
+         */
+        public List<Integer> getState() {
+            return Collections.unmodifiableList(Arrays.asList(stateArray));
+        }
 
-	/**
-	 * 指定されたゴールの要素をGoalの指定された位置(ノード)のスタックにPUSHで
-	 * 設定します。
-	 * @param int index
-	 * @param GoalStackElement elm ゴールの要素
-	 */
-	public void pushGoal(int index, GoalStackElement elm) {
-		goalStackArray[index].add(elm);
-		if(viewer != null) {
-			viewer.repaint();
-		}
-	}
+        /**
+         * 現在の状態をVectorで設定します。
+         * @param Vector state 現在の状態
+         */
+        public void setState(List<Integer> state) {
+            assert state.size() == stateArray.length;
+            for(int i = 0; i < size; i++) {
+                latch.stateArray[i] = state.get(i);
+            }
+            fireSharedMemoryChanged();
+        }
+    }
 
-	/**
-	 * Goalの指定された位置(ノード)の値をスタックから削除します。
-	 * @int index 
-	 */
-	public void removeGoal(int index) {
-		goalStackArray[index].removeLast();
-		if(viewer != null) {
-			viewer.repaint();
-		}
-	}
+    /**
+     * 共有メモリで扱うゴールの情報の単位
+     */
+    public static class GoalStackElement {
+        /**
+         * ゴールの値
+         */
+        public final int value;
+    
+        /**
+         * ゴールを設定したエージェントID
+         */
+        public final int agid;
+    
+        /**
+         * コンストラクタ
+         * @param int value ゴールの値
+         * @param int agid  設定したエージェントのID
+         */
+        public GoalStackElement(int value, int agid) {
+            this.value = value;
+            this.agid = agid;
+        }
+    
+        /**
+         * ゴールの情報を表示します。
+         * @return String ゴールの情報<BR>
+         * 表示形式  val:21 id:701
+         */
+        public String toString() {
+            String str = "val:" + value + " id:" + agid;
+                return str;
+        }
+    
+    }
+    
+    private Latch latch;
 
-	/**
-	 * Goalの要素を全てクリアします。
-	 */
-	public void removeAllGoal() {
-		for(int i = 0; i < LENGTH; i++) {
-			goalStackArray[i].clear();
-		}
-		if(viewer != null) {
-			viewer.repaint();
-		}
-	}
+    private GoalStack goalStack;
 
-	/**
-	 * 現在の状態をVectorで設定します。
-	 * @param Vector state 現在の状態
-	 */
-	public void setState(Vector state) {
-		for(int i = 0; i < LENGTH; i++) {
-			setState(i, (Integer)state.get(i));
-		}
-		if(viewer != null) {
-			viewer.repaint();
-		}
-	}
+    /**
+     * 共有メモリのノード数
+     */
+    private final int size;
+    
 
-	/**
-	 * 全ノードのゴールをVectorで取得します。
-	 * @param Vector GoalValueのVector
-	 */
-	public Vector getGoalValueArray() {
-		Vector goal = new Vector();
-		for(int i = 0; i < LENGTH; i++) {
-			GoalStackElement goalElement = getGoal(i);
-			if(goalElement != null) {
-				goal.add(new Integer(goalElement.value));
-			}else {
-				goal.add(null);
-			}
-		}
-		return goal;
-	}
+    EventPublisherSupport<SharedMemoryEvent, SharedMemoryEventListener> changeListeners = new EventPublisherSupport<SharedMemoryEvent, SharedMemoryEventListener>(SharedMemoryEvent.class, SharedMemoryEventListener.class);
 
+    /**
+     * コンストラクタ
+     * @param int size ノード数（配列のサイズ）
+     */
+    public SharedMemory(int size) {
+        this.size = size;
+        this.latch = new Latch(size);
+        this.goalStack = new GoalStack(size);
+    }
 
-	/**
-	 * Stateの状態を出力します。<BR>
-	 * 出力形式<BR>
-	 * [shared stack]<BR>
-	 *  state<BR>
-	 *   index:0 val:22<BR>
-	 *   index:1 val:15<BR>
-	 *   index:2 val:0<BR>
-	 *   index:3 val:0<BR>
-	 *   index:4 val:1<BR>
-	 */
-	public void printState() {
-		System.out.println("");
-		System.out.println(" state");
-		for(int i = 0; i < LENGTH; i++) {
-			System.out.println("  index:" + i + " val:" + getState(i));
-		}
-	}
+    public GoalStack getGoalStack() {
+        return goalStack;
+    }
 
-	/**
-	 * Goalの状態を出力します。<BR>
-	 * 出力形式<BR>
-	 * [shared stack]<BR>
-	 *  goal<BR>
-	 *   index:0 | 26:200 | 18:100 |<BR>
-	 *   index:1 | 1:200 | 10:100 |<BR>
-	 *   index:2 | 6:100 |<BR>
-	 *   index:3 | 1:1 | 2:101 |<BR>
-	 *   index:4 | 1:101 | 1:101 |<BR>
-	 */
-	public void printGoalStack() {
-		System.out.println("");
-		System.out.println("[shared stack]");
+    public Latch getLatch() {
+        return latch;
+    }
 
-		System.out.println(" goal");
-		for(int i = 0; i < LENGTH; i++) {
-			LinkedList goalStack = goalStackArray[i];
-			int size = goalStack.size();
-			StringBuffer sb = new StringBuffer();
-			sb.append("  index:" + i + " | ");
-			for(int m = 0; m < size; m++) {
-				GoalStackElement elm = (GoalStackElement)goalStack.get(m);
-				if(elm != null) {
-					sb.append(elm.toString() + " | " );
-				}else {
-					sb.append(" no val | ");
-				}
-			}
-			System.out.println(sb.toString());
-		}
-	}
+    /**
+     * ゴールスタックの数を返します
+     */
+    public int getSize() {
+        return size;
+    }
 
+    public void addChangeListener(SharedMemoryEventListener listener) {
+        changeListeners.addEventListener(listener);
+    }
 
+    public void removeChangeListener(SharedMemoryEventListener listener) {
+        changeListeners.removeEventListener(listener);
+    }
+    
+    protected void fireSharedMemoryChanged() {
+        changeListeners.fire("sharedMemoryChanged", new SharedMemoryEvent(this));
+    }
 }
