@@ -1,9 +1,5 @@
 package wba.rogue;
 
-import javax.swing.Timer;
-
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 import java.awt.event.ComponentEvent;
 import java.awt.event.ComponentListener;
 import java.awt.event.KeyEvent;
@@ -13,14 +9,15 @@ import java.awt.Dimension;
 import java.awt.Graphics;
 import java.awt.Graphics2D;
 import java.awt.image.BufferedImage;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
 import javax.swing.JFrame;
 import javax.swing.JPanel;
 import javax.swing.SwingUtilities;
 
 class RoguePanel extends JPanel {
+    /**
+     * 
+     */
+    private static final long serialVersionUID = 1L;
     BufferedImage fb;
     GraphicsRenderer gr;
     final Rogue rogue;
